refactor: move EDID-related utility functions to `edid-field` module
